+++ src/components/SettingsManager/Actions/InputFileSettingControl/index.tsx
@@ -1,7 +1,0 @@
-const InputFieSettingControl = () => {
-  return (
-    <div>InputFieSettingControl</div>
-  )
-}
-
-export default InputFieSettingControl