+++ src/components/SettingsManager/Actions/InputTextSettingControl/index.tsx
@@ -1,9 +1,0 @@
-const InputTextSettingControl = () => {
-  return (
-    <div>InputTextSettingControl
-    
-    </div>
-  )
-}
-
-export default InputTextSettingControl