+++ src/components/SettingsManager/NavigationPanel/index.tsx
@@ -1,7 +1,0 @@
-const NavigationPanel = () => {
-  return (
-    <div>NavigationPanel</div>
-  )
-}
-
-export default NavigationPanel