+++ src/components/SettingsManager/SectionSearchbar/index.tsx
@@ -1,7 +1,0 @@
-const SectionSearchbar = () => {
-  return (
-    <div>SectionSearchbar</div>
-  )
-}
-
-export default SectionSearchbar