+++ src/router.tsx
@@ -1,7 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
 import MainContainer from "./components/MainContainer";
 import Home from './views/Home';
-import Settings from './views/Settings';
 import DownloadFilePopup from './standalone/DownloadFilePopup';
 import StandaloneContainer from './components/StandaloneContainer';
 
@@ -9,7 +8,6 @@ const AppRouter = () => {
     return (
         <Routes>
             <Route path="/" element={<MainContainer><Home /></MainContainer>} />
-            <Route path="/settings" element={<MainContainer><Settings /></MainContainer>} />
             <Route path="/standalone/DownloadFilePopup" element={<StandaloneContainer><DownloadFilePopup /></StandaloneContainer>} />
             <Route path="*" element={<h1>404 Error: Page not found.</h1>} />
         </Routes>