+++ src/views/Settings/index.tsx
@@ -1,9 +1,0 @@
-import SettingsManager from '../../components/SettingsManager'
-
-const Settings = () => {
-  return (
-    <><SettingsManager /></>
-  )
-}
-
-export default Settings